requirements.txt
scapy
dnspython
ipaddress